rename: components의 하위 폴더 이름을 복수형으로 변경
