--- components/atom/index.ts
@@ -1,4 +1,0 @@
-export { default as LikeToggle } from './LikeToggle';
-export { default as LinkButton } from './LinkButton';
-export { default as UserAvatar } from './UserAvatar';
-export { default as NotLoginAvatar } from './NotLoginAvatar';
