+++ components/atoms/index.ts
@@ -1,0 +1,7 @@
+export { default as LikeToggle } from './LikeToggle';
+export { default as LinkButton } from './LinkButton';
+export { default as LinkText } from './LinkText';
+export { default as Logo } from './Logo';
+export { default as Map } from './Map';
+export { default as NotLoginAvatar } from './NotLoginAvatar';
+export { default as UserAvatar } from './UserAvatar';
