+++ components/molecules/Banner/index.tsx
