+++ components/molecules/ExhibitionCard/index.tsx
