+++ components/molecules/ExhibitionCard/style.ts
