+++ components/organisms/CommentList/index.tsx
@@ -4,8 +4,8 @@ import { useInfiniteScroll } from 'hooks';
 import { useEffect, useState } from 'react';
 import Link from 'next/link';
 import { Comment, Tooltip } from 'antd';
-import { UserAvatar } from 'components/atom';
-import { CommentUtils } from 'components/organism';
+import { UserAvatar } from 'components/atoms';
+import { CommentUtils } from 'components/organisms';
 import moment from 'moment';
 import { displayDate } from 'utils';
 import reviewAPI from 'apis/review';
