+++ components/organisms/Footer/style.ts
