+++ pages/reviews/create/index.tsx
@@ -2,8 +2,8 @@ import { useRef, useState, FormEvent, useEffect } from 'react';
 import { reviewAPI } from 'apis';
 import styled from '@emotion/styled';
 import { Input, DatePicker, Switch, Image, Button, message, Form, UploadFile } from 'antd';
-import { Banner } from 'components/molecule';
-import { ImageUpload } from 'components/organism';
+import { Banner } from 'components/molecules';
+import { ImageUpload } from 'components/organisms';
 import { ValueOf } from 'types/utility';
 import { objectToFormData, filesToFormData } from 'utils';
 import imageUrl from 'constants/imageUrl';
