+++ pages/reviews/update/[id]/index.tsx
@@ -2,8 +2,8 @@ import { useRef, useState, FormEvent, useEffect } from 'react';
 import { reviewAPI } from 'apis';
 import styled from '@emotion/styled';
 import { Input, DatePicker, Switch, Image, Button, message, Form, Modal, UploadFile } from 'antd';
-import { Banner } from 'components/molecule';
-import { ImageUpload } from 'components/organism';
+import { Banner } from 'components/molecules';
+import { ImageUpload } from 'components/organisms';
 import { objectToFormData, filesToFormData } from 'utils';
 import { useRouter } from 'next/router';
 import { useAxios } from 'hooks';
