+++ pages/users/[id]/edit/index.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import { Button, Form, Input, Image, message } from 'antd';
 import { useState, useRef, ChangeEvent, FormEvent } from 'react';
-import { SideNavigation } from 'components/molecule';
+import { SideNavigation } from 'components/molecules';
 import { useRecoilValue } from 'recoil';
 import { userAtom } from 'states';
 import { useRouter } from 'next/router';
